+++ shaders/program/composite10.glsl
@@ -12,17 +12,20 @@ const bool colortex0MipmapEnabled = true;
 #endif
 
 void main() {
+	
+	const int bloomIntScale = 1 << BLOOM_RENDER_SCALE;
+	
 	#if HORROR_MODE == 0
-		vec3 bloomColor = texelFetch(MAIN_TEXTURE, texelcoord, 2).rgb * 2.0;
-		float depth = texelFetch(DEPTH_BUFFER_WO_TRANS, texelcoord * 4 + 1, 0).r;
+		vec3 bloomColor = texelFetch(MAIN_TEXTURE, texelcoord, BLOOM_RENDER_SCALE).rgb * 2.0;
+		float depth = texelFetch(DEPTH_BUFFER_WO_TRANS, texelcoord * bloomIntScale + bloomIntScale / 2, 0).r;
 	#else
-		vec3 bloomColor = texture2DLod(MAIN_TEXTURE, texcoord, 2).rgb * 2.0;
+		vec3 bloomColor = texture2DLod(MAIN_TEXTURE, texcoord, BLOOM_RENDER_SCALE).rgb * 2.0;
 		float depth = texture2D(DEPTH_BUFFER_WO_TRANS, texcoord).r;
 	#endif
 	
 	
 	#ifdef DISTANT_HORIZONS
-		float depthDh = texelFetch(DH_DEPTH_BUFFER_WO_TRANS, texelcoord, 0).r;
+		float depthDh = texelFetch(DH_DEPTH_BUFFER_WO_TRANS, texelcoord * bloomIntScale + bloomIntScale / 2, 0).r;
 		float fogAmount = uint(depth == 1.0 && depthDh == 1.0);
 	#elif defined VOXY
 		float fogAmount = uint(depth == 1.0);
@@ -36,19 +39,6 @@ void main() {
 	#endif
 	
 	
-	//const vec3 targetColor = vec3(1.0, 1.0, 0.0);
-	//float bloomMult = dot(bloomColor / getLum(bloomColor), targetColor / getLum(targetColor) * 0.5) * getLum(bloomColor);
-	
-	//float bloomMult = getLum(bloomColor);
-	//bloomMult *= 1.0 + 0.25 * bloomColor.g - bloomColor.r * 0.125;
-	//bloomMult *= bloomMult;
-	
-	//vec3 bloomColorHsv = rgbToHsv(bloomColor + vec3(0.05, 0.05, 0.0));
-	//float bloomMult = 1.0 - 8.0 * abs(bloomColorHsv.x - (51.0 / 360.0));
-	//bloomMult = clamp(bloomMult, 0.0, 1.0);
-	//bloomMult *= bloomColorHsv.z;
-	//bloomMult *= bloomMult;
-	
 	#if HORROR_MODE == 1
 		float bloomMult = dot(bloomColor, vec3(0.2, 0.7, 0.1));
 		bloomMult *= 1.5;
@@ -57,7 +47,6 @@ void main() {
 		bloomMult *= bloomMult;
 	#elif BLOOM_STYLE == 2
 		float bloomMult = getLum(bloomColor);
-		//bloomMult = 1.0 - (1.0 - bloomMult) * (1.0 - bloomMult);
 	#endif
 	
 	bloomMult = smoothstep(BLOOM_LOW_CUTOFF, BLOOM_HIGH_CUTOFF, bloomMult);