+++ shaders/style_cartoon.glsl
@@ -526,6 +526,10 @@
 
 
 
+#if BLOOM_STYLE == -1
+#undef BLOOM_STYLE
+										#define BLOOM_STYLE 1
+#endif
 #if BLOOM_AMOUNT == -1
 #undef BLOOM_AMOUNT
 										#define BLOOM_AMOUNT 0.9
@@ -549,10 +553,6 @@
 #if BLOOM_HIGH_CUTOFF == -1
 #undef BLOOM_HIGH_CUTOFF
 										#define BLOOM_HIGH_CUTOFF 1.0
-#endif
-#if BLOOM_STYLE == -1
-#undef BLOOM_STYLE
-										#define BLOOM_STYLE 1
 #endif